requirements.txt
Django==4.2.5
django-cors-headers==4.3.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
whitenoise==6.6.0
python-Levenshtein==0.25.1
PyPDF2==3.0.1